Create 8-10-17-Regrid-area-adj-data
This can calculate the adjacencies for each cell, as before. Now it can also write the file that says which cells are next to which gauges and structures. Also re-calculates the area of each new grid that is larger than 400m by counting how many original grids are in it.