Create 1-24-18-newdsmax-d.dry
added separate ds.max statements for when d.stage is too positive and too negative. also added statements to prevent stage from going too low. when set to ground level, the model was not able to go down to the lower third of the stage record, so I added d.dry, the distance below ground that the model was allowed to go (0.4m). Issue will not be solved until I find the missing. 0.25m of water.